Make small changes.
