+++ batchglm/train/numpy/glm_poisson/model_container.py
@@ -5,16 +5,7 @@ import numpy as np
 import scipy
 
 from .external import BaseModelContainer
-
-
-class NoScaleError(Exception):
-    """
-    Exception raised for attempting to access the scale parameter (or one of its derived methods) of a poisson model.
-    """
-
-    def __init__(self, method):
-        self.message = f"Attempted to access {method}. No scale parameter is fit for poisson - please use location."
-        super().__init__(self.message)
+from .exceptions import NoScaleError
 
 
 class ModelContainer(BaseModelContainer):
